Use Decimal, not float for currencies.
A colleague rightly suggested that floats are dangerous for currency
because of their rounding effects. Even if, the lib did not do any
mathematical operations on this data and the risk did not exist, floats
lowered confidence an end user might have in using the library.

Also modify json serializer to cast back to string. Using custom
JsonEncoder was thought through and resisted. The use case is small enough.